Consolidate database migrations
