+++ src/database/__init__.py
@@ -1,0 +1,40 @@
+"""
+Copyright (C) 2024 Michael Piazza
+
+This file is part of Smart Notes.
+
+Smart Notes is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+Smart Notes is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with Smart Notes.  If not, see <https://www.gnu.org/licenses/>.
+"""
+
+from .connection import (
+    get_database_path,
+    get_sqlite_backend,
+    get_user_files_path,
+    open_database,
+)
+from .migrations import (
+    apply_database_bootstrap_migrations,
+    apply_database_migrations,
+    run_migrations,
+)
+
+__all__ = [
+    "apply_database_bootstrap_migrations",
+    "apply_database_migrations",
+    "get_database_path",
+    "get_sqlite_backend",
+    "get_user_files_path",
+    "open_database",
+    "run_migrations",
+]
