+++ src/database/db_migrations/0001_initial_smart_fields_schema.py
