+++ src/database/db_migrations/0002_migrate_deprecated_chat_models_to_auto.py
