+++ src/database/migrations.py
@@ -17,42 +17,44 @@ You should have received a copy of the GNU General Public License
 along with Smart Notes.  If not, see <https://www.gnu.org/licenses/>.
 """
 
-import sqlite3
 from pathlib import Path
 from typing import Optional
 
 from yoyo import read_migrations
-from yoyo.backends.core.sqlite3 import SQLiteBackend
-from yoyo.connections import default_migration_table, parse_uri
 
-from .logger import logger
+from ..logger import logger
+from . import connection
+from .legacy_config_migration import migrate_legacy_config_to_database
 
-DATABASE_FILENAME = "smart_notes.sqlite3"
-USER_FILES_DIR = "user_files"
 
-# Bootstrap migrations create only the database shape needed to import legacy
-# config-backed Smart Fields. Data migrations must run after that import so
-# rows copied from old config are included.
+def run_migrations() -> None:
+    # Bootstrap creates the schema only. Generation defaults and Smart Fields
+    # both come from legacy config.json, so import both before SQL data
+    # migrations. That keeps future model backfills simple: update the default
+    # row plus custom override rows, and inherited fields follow automatically.
+    apply_database_bootstrap_migrations()
+    migrate_legacy_config_to_database()
+    apply_database_migrations()
 
 
 def apply_database_bootstrap_migrations(database_path: Optional[str] = None) -> None:
-    apply_migrations(database_path, bootstrap_only=True)
+    _apply_migrations(database_path, bootstrap_only=True)
 
 
 def apply_database_migrations(database_path: Optional[str] = None) -> None:
-    apply_migrations(database_path)
+    _apply_migrations(database_path)
 
 
-def apply_migrations(
+def _apply_migrations(
     database_path: Optional[str] = None,
     bootstrap_only: bool = False,
 ) -> None:
     # Tests pass isolated temp DB paths so migration state never touches user data.
-    resolved_database_path = database_path or get_database_path()
+    resolved_database_path = database_path or connection.get_database_path()
     Path(resolved_database_path).parent.mkdir(parents=True, exist_ok=True)
 
     logger.debug(f"Smart fields DB: preparing migrations for {resolved_database_path}")
-    backend = get_sqlite_backend(resolved_database_path)
+    backend = connection.get_sqlite_backend(resolved_database_path)
     migrations_path = Path(__file__).with_name("db_migrations")
     logger.debug(f"Smart fields DB: reading migrations from {migrations_path}")
     migrations = read_migrations(str(migrations_path))
@@ -70,29 +72,3 @@ def apply_migrations(
         )
         backend.apply_migrations(pending_migrations)
         logger.info("Smart fields DB: database migrations applied")
-
-
-def open_database(database_path: Optional[str] = None) -> sqlite3.Connection:
-    resolved_database_path = database_path or get_database_path()
-
-    conn = sqlite3.connect(resolved_database_path)
-    conn.row_factory = sqlite3.Row
-    conn.execute("PRAGMA foreign_keys = ON")
-    return conn
-
-
-def get_database_path() -> str:
-    return get_user_files_path(DATABASE_FILENAME)
-
-
-def get_user_files_path(filename: str) -> str:
-    return str(Path(__file__).resolve().parent.parent / USER_FILES_DIR / filename)
-
-
-def get_sqlite_backend(database_path: str) -> SQLiteBackend:
-    # In Anki's vendored add-on environment, yoyo's package metadata may be missing.
-    # Import the SQLite backend directly instead of relying on entry-point discovery.
-    database_uri = f"sqlite:///{Path(database_path).absolute().as_posix()}"
-    backend = SQLiteBackend(parse_uri(database_uri), default_migration_table)
-    backend.init_database()
-    return backend
