+++ tests/test_smart_field_prompt_map.py
@@ -39,9 +39,10 @@ DECK_ID = cast(DeckId, 1)
 @pytest.fixture(autouse=True)
 def sqlite_database(tmp_path: Path, monkeypatch: pytest.MonkeyPatch) -> None:
     import src.database
+    import src.database.connection
 
     monkeypatch.setattr(
-        src.database,
+        src.database.connection,
         "get_database_path",
         lambda: str(tmp_path / "smart_notes.sqlite3"),
     )
